adapt main to changes in actors, delete server mod and unused fx
